mediaverse_BE/requirements.txt
django
djangorestframework
psycopg2
pillow
drf-spectacular
django-cors-headers
